Project 1
Simple game created.

Goal is to take two players (1 & 2), and have a random dice roll. When player 1 rolls an odd number or player 2 rolls an even they win and gain points. However, when player 1 rolls an even number or player 2 rolls an odd number they lose points. These points are randomly generated and changed due to game rules that are made. REGULAR does nothing to the points, DOUBLE doubles the points, and BONUS sets the points strictly to 200 for that round. The user inputs how many rounds are to be played. Once the number of rounds inputted by the user are played the player with the highest score wins!!!